requirements.txt
torch
torchvision
matplotlib
tqdm
Pillow
albumentations
numpy
polib  #My personal lib 
